Big changes
Changed the functions for fetching data from the API to the modern style using async and await.

Changed APIs to openweathermap.

Added functions to update the frontend HTML with the fetched data from the API.

Added a search button for the user to click or press enter to look for a city.

Added styles for body, fonts, cards, divs, buttons, search bars, etc.

Changed the HTML code to a simpler design which is to be expanded on later.