Added Settings Menu
Added settings menu with checkboxes to toggle on and off certain attributes like feelsLike temperature, wind direction, etc.

Added a button to go toggle the settings menu.

Added a back button to return to the home page.

Added localStorage to store user data in order to only display the attributes that the user wants.